requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.38.0
Django==4.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
filetype==1.2.0
idna==3.6
importlib-resources==6.1.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
zipp==3.17.0
